Catch data TN

Defines the topological sequence of catchments for nitrogen.

catch_data_TN

Format

A data frame with 5 variables:

HydroID

integer positive. Unique catchment identifier.

To_catch

integer. Unique identifier of the catchment to which the catchment goes. Note that for the outlet To_catch== -1.

Shreve

integer. this indicates the Shreve order of the topological sequence in the stream network.

LakeFrRet

fraction, 0-1. Lake retention fraction.

NrmLengthKm

double. Normalized length of catchment reach.