Evolution plot area

This function blah, blah, blah....

evolution_plot_area(
  annual_data,
  data,
  title = NULL,
  xaxis.title = NULL,
  yaxis.title,
  x,
  y,
  colour = NULL,
  wrap = NULL
)

Arguments

annual_data, data

data frame.

title

character. The plot title

xaxis.title

character. The title of x axis

yaxis.title

character. The title of y axis

x

character. The name of variable in data to represent in x axis

y

character. The name of variable in data to represent in y axis

colour

character. The name of variable in data to represent colour

wrap

character. The name of variable in data to represent wrap