Nutrient balance based in the application of the Geospatial Regression Equation returning the diffuse, land retention, point sources

Computes the basin nutrient balance.

nut_balance(
  catch_data,
  annual_data,
  alpha_p,
  alpha_l,
  sd_coef,
  loc_years,
  atm_coeff = 0.38
)

Arguments

catch_data

data frame. Definition of the topological sequence of catchments.

annual_data

data frame. Sources of nutrient for each year and catchments.

alpha_p

numeric. First model parameter, the basin retention coefficient.

alpha_l

numeric. Second model parameter, the river retention coefficient.

sd_coef

numeric. Third model parameter, fraction of domestic diffuse sources that reaches the stream network.

loc_years

integer. Years in which the model should be executed.

atm_coeff

numeric. A value for atmospheric attenuation coefficient.

Value

One object, a data frame

Examples

# the data of the TN scenario
data(catch_data_TN)
data(annual_data_TN)
# the parameter to assess the basin model
alpha_p <- 35.09
alpha_l <- 0.02
sd_coef <- 0.2
# year in which the model should be executed
loc_years <- 1990:2018
# Computing the nutrient balance
basin_loads_b <- nut_balance(catch_data_TN, annual_data_TN, alpha_p, alpha_l,
sd_coef, loc_years)
#> [1] "Calculating year: 1990"
#> [1] "Calculating year: 1991"
#> [1] "Calculating year: 1992"
#> [1] "Calculating year: 1993"
#> [1] "Calculating year: 1994"
#> [1] "Calculating year: 1995"
#> [1] "Calculating year: 1996"
#> [1] "Calculating year: 1997"
#> [1] "Calculating year: 1998"
#> [1] "Calculating year: 1999"
#> [1] "Calculating year: 2000"
#> [1] "Calculating year: 2001"
#> [1] "Calculating year: 2002"
#> [1] "Calculating year: 2003"
#> [1] "Calculating year: 2004"
#> [1] "Calculating year: 2005"
#> [1] "Calculating year: 2006"
#> [1] "Calculating year: 2007"
#> [1] "Calculating year: 2008"
#> [1] "Calculating year: 2009"
#> [1] "Calculating year: 2010"
#> [1] "Calculating year: 2011"
#> [1] "Calculating year: 2012"
#> [1] "Calculating year: 2013"
#> [1] "Calculating year: 2014"
#> [1] "Calculating year: 2015"
#> [1] "Calculating year: 2016"
#> [1] "Calculating year: 2017"
#> [1] "Calculating year: 2018"